refactor(classroom): Separate NoClassCodeComponent from DashboardScreen (#20)
